+++ test/test_inkpen.rb
@@ -7,7 +7,19 @@ class TestInkpen < Minitest::Test
     refute_nil ::Inkpen::VERSION
   end
 
-  def test_it_does_something_useful
-    assert false
+  def test_configuration_exists
+    refute_nil Inkpen.configuration
+  end
+
+  def test_configure_block_works
+    Inkpen.configure do |config|
+      assert_kind_of Inkpen::Configuration, config
+    end
+  end
+
+  def test_reset_configuration
+    original = Inkpen.configuration
+    Inkpen.reset_configuration!
+    refute_same original, Inkpen.configuration
   end
 end